+++ src/StartupManager/Commands/StartupList/ListPrograms.cs
@@ -8,16 +8,18 @@ namespace StartupManager.Commands.StartupList {
         public bool Disabled { get; set; }
         public bool AllUsers { get; set; }
         public string RegistryPath { get; set; }
+        public string DisabledRegistryPath { get; set; }
         public string RegistryName { get; set; }
         public StartupType Type { get; set; }
 
-        public ListProgram(string name, [AllowNull] string path, bool requireAdministrator, bool disabled, StartupType type, bool allUsers, string registryPath = "", string registryName = "") {
+        public ListProgram(string name, [AllowNull] string path, bool requireAdministrator, bool disabled, StartupType type, bool allUsers, string registryPath = "", string disabledRegistryPath = "", string registryName = "") {
             Path = path ?? string.Empty;
             RequireAdministrator = requireAdministrator;
             Disabled = disabled;
             Type = type;
             AllUsers = allUsers;
             RegistryPath = registryPath;
+            DisabledRegistryPath = disabledRegistryPath;
             RegistryName = registryName;
             Name = ParseName(name);
         }