+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2019-11-08
+
+### Added
+
+- #2 It's now possible to remove programs from starting with windows.
+
 ### Fixed
 
 - List output had too many newlines as seperator between header.
@@ -65,7 +71,8 @@ No new application features, but streamlined build, publish and release process.
 - Enable/Disable command
   - It's now possible to enable/disable existing programs starting with windows.
 
-[Unreleased]: https://github.com/Faustvii/StartupManager/compare/1.3.0...HEAD
+[Unreleased]: https://github.com/Faustvii/StartupManager/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/Faustvii/StartupManager/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/Faustvii/StartupManager/compare/1.2.1...1.3.0
 [1.2.1]: https://github.com/Faustvii/StartupManager/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/Faustvii/StartupManager/compare/1.1.0...1.2.0